adding workspace for notebook
